- App has been moved to root compoents - Add Loader as a Component
