+++ src/components/Loader/Loader.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import styles from './Loader.scss';
+
+const Loader = () => {
+  return (
+    <div className="loader">Loading</div>
+  );
+}
+
+export default Loader;
